requirements.txt
cffi==1.5.2
characteristic==14.3.0
croniter==0.3.11
cryptography==1.2.3
decorator==4.0.9
enum34==1.1.2
idna==2.0
ipaddress==1.0.16
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.14
pyOpenSSL==0.15.1
python-dateutil==2.5.0
service-identity==16.0.0
six==1.10.0
Twisted==15.5.0
wokkel==0.7.1
zope.interface==4.1.3
